Add fallback to getInitials function and remove duplicate
